ModelMapper config and DTO implementation
The Category DTO only takes the id and the description of the Category Entity, since the list of notes is being fetched separately from the client. The Note DTO takes the same attributes of the Note Entity. The ModelMapper is initialized as a Bean in MapperConfig.java and injected in the service layer for each entity